Move Akai Fire controller into akai package
